Reworked Storyboard Flow; Added "Manage Child Profiles" and "Manage Subscription";
